archive/requirements.txt
Flask
gevent
gevent-websocket
flask_sockets
openai-whisper
werkzeug
numpy